% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_4chan_thread.R
\name{get_4chan_thread}
\alias{get_4chan_thread}
\title{Return specific 4chan thread}
\usage{
get_4chan_thread(board, thread_id)
}
\arguments{
\item{board}{Character variable of the 4chan board.\cr
Available boards are: "adv", "plebs", "hr", "tg", "tv", "x", "s4s", "pol",
"o", "trv", "f", "sp", "mlpol", "mo".}

\item{thread_id}{Dataframe with details of all posts in the thread.}
}
\value{
Dataframe with details on all posts in the thread.
}
\description{
Return 4chan thread by looking up its ID
}
\details{
Variables in API output:\cr\cr
thread_id: 4chan ID of the thread the post is situated in\cr
doc_id: 4chan document ID\cr
num: 4chan post ID\cr
subnum: Binary, 1: Ghost Post, 0: Non-Ghost Post\cr
op: Binary, 1: Opening Post, 0: Reply Post\cr
timestamp: Time sent in Posix time\cr
fourchan_date: Time sent\cr
name: Author name\cr
title: Post title\cr
referencing_comment: The post_id this post is quoting\cr
comments: Text of the post
poster_country: Author country\cr
nreplies: Number of replies\cr
formatted: Boolean, Has this post been formatted?\cr
media_link: Download link to the media (e.g. images) that have been shared in
the post
}
\examples{
\dontrun{
get_4chan_thread(board = "adv", thread_id = 21738271)

get_4chan_thread(board = "trv", thread_id = 1888747)
}
}
\seealso{
\code{\link[httr]{modify_url}}, \code{\link[httr]{user_agent}},
\code{\link[httr]{GET}}, \code{\link[httr]{http_type}},
\code{\link[httr]{content}}, \code{\link[httr]{http_error}}
\code{\link[jsonlite]{toJSON, fromJSON}}
\code{\link[purrr]{map}}, \code{\link[purrr]{map2}}
\code{\link[stringr]{str_extract}}
}
