% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_4chan.R
\name{search_4chan}
\alias{search_4chan}
\title{Search all 4chan posts}
\usage{
search_4chan(
  start_date = "",
  end_date = "",
  boards,
  text = "",
  subject = "",
  filename = "",
  ghost = "all",
  user_id = "",
  tripcode = "",
  type = "all",
  username = "",
  country = "",
  results = "all",
  show_only = "all",
  deleted = "all",
  capcode = "user",
  order = "asc",
  cool = 20
)
}
\arguments{
\item{start_date}{String, Start date of post search. Format: "YYY-MM-DD",
Default: ''}

\item{end_date}{String, End date of post search. Format: "YYY-MM-DD",
Default: ''}

\item{boards}{String or string vector of the 4chan board.\cr
Available boards are: "adv", "plebs", "hr", "tg", "tv", "x", "s4s", "pol",
"o", "trv", "f", "sp", "mlpol", "mo".}

\item{text}{String, Searchterms in post text. Regex possible, Default: ''}

\item{subject}{String, Searchterms in post subject. Regex possible,
Default: ''}

\item{filename}{String, Searchterms in filename. Regex possible, Default: ''}

\item{ghost}{String, Filter by: "all", "only", "none" , Default: 'all'}

\item{user_id}{Integer, Filter by user_id, Default: ''}

\item{tripcode}{String, Filter by tripcode, Default: ''}

\item{type}{String, Filter by post type: "all", "sticky", "op", "posts",
Default: 'all'}

\item{username}{String, Search username, Default: ''}

\item{country}{String, Filter by country. Format:
ISO 3166-1 Alpha-2 country code, Default: ''}

\item{results}{String, Return post matches or the thread they are found in.
Possible values: "all", "thread", Default: 'all'}

\item{show_only}{String, Show only posts containing data types: "all",
"text", "image", Default: 'all'}

\item{deleted}{String, Filter deletion status: "all", "not-deleted",
"deleted", Default: 'all'}

\item{capcode}{String, Filter author type: "all", "user", "ver", "mod",
"dev", "admin", "manager", "founder", Default: 'all'}

\item{order}{String, Return posts ordered by latest or oldest: "asc",
"desc", Default: 'asc'}

\item{cool}{Integer (seconds), The 4plebs API includes an API rate limit
of 5 requests/min for the archive search. For multiple searches a cool-down
is recommended , Default: 20}
}
\value{
Dataframe with details of the posts found.
}
\description{
This function returns all 4chan posts specified by searchterms
}
\details{
Variables in API output:\cr\cr
thread_id: 4chan ID of the thread the post is situated in\cr
doc_id: 4chan document ID\cr
num: 4chan post ID \cr
subnum: Binary, 1: Ghost Post, 0: Non-Ghost Post\cr
op: Binary, 1: Opening Post, 0: Reply Post\cr
timestamp: Time sent in Posix time\cr
fourchan_date: Time sent\cr
name: Author name\cr
title: Post title\cr
referencing_comment: The post_id this post is quoting\cr
comments: Text of the post\cr
poster_country: Author country\cr
nreplies: Number of replies\cr
formatted: Boolean, Has this post been formatted?\cr
media_link: Download link to the media (e.g. images) that have been shared
in the post
}
\examples{
\dontrun{
search_4chan(boards = "trv", start_date = "2021-04-20",
    end_date = "2022-12-21", text = "mallorca|menorca")

search_4chan(boards = "o", text = "simson", show_only = "image")
}
}
