% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_4chan_board_index.R
\name{get_4chan_board_index}
\alias{get_4chan_board_index}
\title{Return threads for specific 4chan board index}
\usage{
get_4chan_board_index(board, page, latest_comments = F, cool = 0)
}
\arguments{
\item{board}{Character variable of the 4chan board.\cr
Available boards are: "adv", "plebs", "hr", "tg", "tv", "x", "s4s", "pol", "o", "trv", "f", "sp", "mlpol", "mo".}

\item{page}{Integer of the board page index}

\item{latest_comments}{Boolean, TRUE: Return opening posts and all replies, FALSE: Return only opening posts, Default: FALSE}

\item{cool}{Integer (seconds), The 4plebs API includes an undocumented API rate limit for the board index search. For multiple searches a cool-down is recommended , Default: 0}
}
\value{
Dataframe with details on all posts on a given board page.
}
\description{
This function returns all 10 threads from a given board page index, starting from the latest thread
}
\details{
Variables in API output:
thread_id: 4chan ID of the thread the post is situated in
doc_id: 4chan document ID
num: 4chan post ID
subnum: Binary, 1: Ghost Post, 0: Non-Ghost Post
op: Binary, 1: Opening Post, 0: Reply Post
timestamp: Time sent in Posix time
fourchan_date: Time sent
name: Author name
title: Post title
referencing_comment: The post_id this post is quoting
comments: Text of the post
poster_country: Author country
nreplies: Number of replies
formatted: Boolean, Has this post been formatted?
media_link: Download link to the media (e.g. images) that have been shared in the post
}
\examples{
\dontrun{
get_4chan_board_index(board = "mo", page = 1, latest_comments = TRUE)

get_4chan_board_index(board = "mo", page = 1, latest_comments = FALSE)
}
}
\seealso{
\code{\link[httr]{modify_url}}, \code{\link[httr]{user_agent}}, \code{\link[httr]{GET}}, \code{\link[httr]{http_type}}, \code{\link[httr]{content}}, \code{\link[httr]{http_error}}
 \code{\link[jsonlite]{toJSON, fromJSON}}
 \code{\link[purrr]{map}}, \code{\link[purrr]{map2}}
 \code{\link[stringr]{str_extract}}
}
